Improve burst support, move the namespace
Adds optional parameters size and position that are used to determine
the final size and position of a burst template. Adds a lot of fine
tuning for template sizing on the burst option for best conformance to
the rulebook, but very large bursts won't line up super well onless the
template shape can be changed to a bounding hex.

Future work may remove the new properties in favor of a token aware
snapping function.

Moves the namespace from game.lancer.entities to game.lancer.canvas.