Make Burst templates moveable.
Burst templates no longer need optional parameters, as they now size and
snap based on hovered tokens. The actual logic for how to snap is
slightly nightmarish due to the fact that on hex grids, tokens
frequently overlap. The logic used is to snap to the token center that
is nearest to the cursor. Complexity is O(n) for number of tokens on the
canvas.