Add information to the flags of templates
Stores the range and creator (token id) used to create the template.
Additionally stores a list of token ids to ignore for targeting (default
only contains the creator) and a list of token dispositions to ignore
(default empty).

These flags can then be manipulated before calling drawPreview and the
final created template will be created with those flags intact.