Use _preCreate method over Hooks
_preCreate is a new addition in 0.8 and has a couple of benefits over
the hook. Biggest is that it doesn't rely on setting up a Hook in
lancer.ts. Another advantage is that it is guaranteed to run and finish
before and preCreateDocument hook that modules might be providing.
Another big one is that the type data from the parent Document can be
used.