Add a button for deploying a WeaponRangeTemplate to the accdiff dialog
The button only shows for weapons that have a range listed that is an
AoE range (such as Burst 3) and clicking it fades the HUD out until
placement is either completed or cancelled. Upon placement, units that
are within the highlighted area of the MeasuredTemplate are set as the
user's targets and the dialog refreshes with the new targets.