Set prototype token size based on unit size.
Automatically sets the size of prototype tokens to match the size of the
actor, as computed by machine-mind. Manual changes will be persisted
until the mm computed size changes. Manually changing the size of the
Frame, NpcClass or Deployable will immediately change the prototype as
will swapping Class or Frame.

Additionally sets imported mechs to the same disposition as their pilot
and deployables to neutral, as well as setting actorLink for mechs.