Add getCritRoll function to macros
In order to reuse the dice from a normal hit in a critical hit, we need
to manually construct a new roll using the original one. The function
takes a roll, then makes a second roll with the same formula, manually
smooshes the two rolls into one with the appropriate dice kept, then
returns the new roll, which is already evaluated.