Await all crit rolls, fix has_normal_hit, fix kept dice in getCritRoll
Since getCritRoll is async it ought to be awaited. Wraps the map in an
awaited Promise.all so that they can at least run in parallel.

Puts parentheses around `(attack.roll.total ?? 0)` so that it actually
evaluates to false if its a crit on an untargeted roll.

Checks for the existence of a kh modifier and uses that before falling
back on the number of dice for the term. This prevents 0 dice being
kept active when there's already a kh clause on the roll.