Fix the cache invalidation strategy for getOccupiedSpaces()
Since _onUpdate seems to run before prepareData, and vision is
calculated before prepareData, sensor-sight exposed a race condition
that would cause tokens to cache their old occupied spaces when moving.
This now stores the floored coordinate and compares it the the current
coord and invalidates the cache if they differ in the getOccupiedSpaces
method directly.