Check for the bars container before redrawing token bars
Foundry doesn't initialize the token components until the draw method is
called. This leads to a race condition where it is possible to drop a
token, then attempt to open the sheet, leading to drawBars to be called
before the canvas has drawn. This leads to sheets becoming un-openable
until a reload or update to the actor.

This contains a compatibility path for changes to tokens in v9 and will
work on both 0.8.x and v9.