Update lancer-initiative to take advantage of v9 features
Updates lancer initiative to the v9 only 2.0.1 which no longer uses a
dummy combatant. This could potentially improve compatibility with
modules that incorrectly assume that actors and tokens are required to
be associated with combatants.

Pulls in changes to the combat tracker template from core.

This also marks the system as _only_ v9 compatible (which it already
was).