Update handleCombatUpdate for lancer-initiative 2.x
The function handleCombatUpdate was written with the assumption that
turn is never null, however in Foundry v9, turn can be set to null to
indicate no turn. The lancer-initiative library was updated to take
advantage of this, which causes errors when changing rounds, ending
turns, or selecting a turn when there's no acive combatant. Now, check
whether a referenced turn is null, and if so, skip tryitn to process
that turn.