Fix NPC class/template feature removal behavior
Before, when removing NPC features from NPC classes or templates, the
system would attempt to delete the feature. This is correct behavior for
Actors, but when this happens on an Item (e.g. NPC Class or Template),
this would either attempt to delete the feature from the Compendium
(throwing an error since the compendium is locked), or if the feature
was a custom feature added by the GM, the feature would be fully
deleted from the world and item list.

This commit changes removal behavior when the current sheet is an item
sheet, in which case, instead of calling destroy_entry(), the item is
spliced out of the correct list on the sheet (i.e. base features and
optional features).

Fixes #435