Use JQuery to select element and add handler
If more than one sheet is open, the handlers are applied to only one
sheet. In that case, uploading to that sheet will update all open sheets
and uploading to a sheet that didn't recieve its handler will do
nothing. By using the html parameter, only the input on the current
sheet will be found and given the event handler.