Update compcon-login-form.ts
I submitted a bug mentioning the lowercase email issue; here is my simple proposed solution. Essentially, YOU make the email lowercase just before sending the form data through the network; that way the users can type in whatever casing they want and it won't be an issue.

I assume it was compcon that was forcing all emails to be lowercase, this'll make your app a little more case friendly.