Update esmodule entry point.
With the vite upgrade, whatever was causing the entry point to be
generated as index.js instead of lancer.js was, resolved. Additionally,
vite now uses a mjs extension for generated esmodules.

This oversight was due to the change to the vite config to no empty the
build directory masking the change by leaving the old index.js intact.