Glitchtext for scanning Abominable characters to journal
Adds functionality to the Scan (Journal) macro which enables it to render glitchy, distorted unreadable garbage in place of the normal results when the scan target is Abominable (from the Horror template from Dustgrave).

Some information is still conveyed - the number of templates, the number of systems from each one, the number of weapons, and of course that it is a Horror and it has the Abominable feature. Additionally, the glitchy text roughly matches the length of the original text. 

If Abominable is marked as destroyed, the scan will work as normal.