Rewrite LancerToken#getOccupiedSpaces
This refactors getOccupiedSpaces to use cubic coordinates and filtering
to determine the needed spaces on hex grids. Also refactors the square
calculation also avoiding PIXI. Since this approach doesn't need to use
PIXI to test if the points are within the border polygon, it is
significantly faster. Additionally, gridless tokens will now return
their center point as the sole element of the array, though smarter
calculations are still recommended.