Create fromLidMany helper
This uses the ability to pass a query to getDocuments in order to
retrieve documents from an array of lids without looping. Also refactors
fromLid to retrieve using fromLidMany and simply return the first
document (if found).

While performance should be roughly the same for the single document
case, this should allow for marginally faster retrievals of multiple
documents by reducing the number of database requests.