Add unused attack dialog tour
For technical reasons, the attack tour is highly unreliable. Particularly,
individual animation cannot be awaited, and seem to be quite long
requiring excessive delaying between steps. Furthermore, due to the
hookup of keyboard handlers on hud open, it's possible to dismiss the
HUD without ending the tour, either by confirming it (which logs an
error from the attack flow), or by tapping escape which should dismiss
the tour, but the dialog eats the esc. Finally foundry has difficulty
determining the location of the element of the targeting step, and the
tooltip ends up on top of the element.