fix: players can reset actions w/o "Allow Players"
Fixes a bug where players could reset their actions (via button on Action Tracker) when the `Allow Players` action tracker setting is disabled.

If players are not allowed to use action buttons to modify their actions, they should not be able to reset their actions either.

In addition to guarding against this, this also only renders the reset button when the user is allowed to use it.