Begin style reorg
The styles are a huge mess and getting pretty close to unmaintainable.
This doesn't really adress the big issues, but it adds in some an
example of how to organize the styles going forward. Long term,
eliminating top-level classes and using @import for reuse should be the
goal. They aren't touched here, but :global rules in svelte files should
be either moved to an actual global space or converted to an importable
file, whether a plain include or a library.