fix: expose LancerToken hover methods
Also refactor highlightAttackAndDamageTargets to use native HTML. This
might be a breaking change in terms of the use markup. I looked at the
templates and the UUID for the target was directly on the element, the
listener is registered on, so removed the `closest` element query. If
the UUID is not always on that element, that query needs to be put back
in.